--- frontend/src/components/Pages/SignInPage/SignInButton.tsx
@@ -1,23 +1,0 @@
-import { useMsal } from '@azure/msal-react'
-import { loginRequest } from '../../../api/AuthConfig'
-import { Button } from '@equinor/eds-core-react'
-import { IPublicClientApplication } from '@azure/msal-browser'
-
-function handleLogin(instance: IPublicClientApplication) {
-    instance.loginRedirect(loginRequest).catch((e) => {
-        console.error(e)
-    })
-}
-
-/**
- * Renders a button which, when selected, will redirect the page to the login prompt
- */
-export const SignInButton = () => {
-    const { instance } = useMsal()
-
-    return (
-        <Button href="" variant="contained" onClick={() => handleLogin(instance)}>
-            Sign in using Redirect
-        </Button>
-    )
-}
