+++ frontend/src/components/Pages/SignInPage/SignInPage.tsx
@@ -1,5 +1,27 @@
 import { useIsAuthenticated } from '@azure/msal-react'
-import { SignInButton } from './SignInButton'
+import { useMsal } from '@azure/msal-react'
+import { loginRequest } from '../../../api/AuthConfig'
+import { Button } from '@equinor/eds-core-react'
+import { IPublicClientApplication } from '@azure/msal-browser'
+
+function handleLogin(instance: IPublicClientApplication) {
+    instance.loginRedirect(loginRequest).catch((e) => {
+        console.error(e)
+    })
+}
+
+/**
+ * Renders a button which, when selected, will redirect the page to the login prompt
+ */
+export const SignInButton = () => {
+    const { instance } = useMsal()
+
+    return (
+        <Button href="" variant="contained" onClick={() => handleLogin(instance)}>
+            Sign in using Redirect
+        </Button>
+    )
+}
 
 export const SignInPage = () => {
     const isAuthenticated = useIsAuthenticated()
